requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
faiss-cpu
chromadb
tiktoken
aiofiles
python-multipart
qdrant-client
python-dotenv

